Rename splash/presentation to views
