+++ lib/main.dart
@@ -3,7 +3,7 @@ import 'package:librarian_app/constants.dart';
 import 'package:librarian_app/src/features/borrowers/data/borrowers_view_model.dart';
 import 'package:librarian_app/src/features/loans/data/loans_view_model.dart';
 import 'package:librarian_app/src/features/authentication/data/user_view_model.dart';
-import 'package:librarian_app/src/features/splash/presentation/splash_page.dart';
+import 'package:librarian_app/src/features/splash/views/splash_page.dart';
 import 'package:provider/provider.dart';
 import 'package:supabase_flutter/supabase_flutter.dart' as supabase;
