+++ lib/src/features/splash/views/splash_page.dart
